aufs: intro, public header
A header file for both of kernelspace and userspace.

For the new file fs/aufs/Kconfig, the maximum number of branches is
customizable, and it determines the type (size) of 'aufs_bindex_t.' The
type is always 'signed.' If we made it 'unsigned,' then more branches
would be available. But generally I think 127 (default) is enough and it
won't be a big issue.

For those who wants more than 127 branches, other values are
available. But we should care the size of the internal pointer arrays,
and it is good for the performance to keep it in a page at most.
AUFS_BRANCH_MAX_511 is mainly for 64bit systems which limits the
internal array size less than 4k (511 x 8bytes < 4k). Similarly for
32bit systems, AUFS_BRANCH_MAX_1023 (1023 x 4 bytes < 4k).

See also the documents in this commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>