aufs: sbinfo list
When user accesses aufs via other than fs related systemcalls, aufs
needs to identify which superblock is the target.  Here is the trick.
It is just a list of aufs superblocks.  Such way will be procfs and
MagicSysRq key.  For MagicSysRq support, see the later commit.
This is a dirty approach which I don't like, but I just don't have
another idea.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>