aufs: virtual or vertical directory 2/2, body
It is hard to implement readdir(3) for aufs virtual directory.
It considers the every whiteout in a single direcotry, as well as the
(first) opaque marker (diropq).
This implementation consumes memory a lot, and I'd suggest you to try
RDU (readdir in userspace) in later commit.

See also the document in previous commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>