aufs: support FUSE branch a little more
Generally FUSE creates a new inode object when the internal field of
it is modified.  Additionally some FUSE-based fs don't maintain the
inode's attributes such as permission bits and link count.  It is hard
for aufs to live with such FS and calls BUG().
But in some cases, aufs should not be so strict since FUSE hides the
problem from the users' point of view.
Testing Mauricio's ntfs-3g case, aufs decided to relax the condition
of BUG() for some FUSE branch.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>