aufs: fs_context 1/7, skelton of the new shceme
Introduces a new file fs/aufs/fsctx.c which defines
- file_system_type->init_fs_context()
- file_system_type->parameters
- fs_context_operations->free()
but they are not completed and not activated yet.
This series "fs_context" will add the support for fs_context step by
step, and these operations will be filled and set in the succeeding
commits.

The commit itself is rather meaningless and never affect the
behaviour.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>