Revert "aufs: bugfix, copy FMODE_CAN_ODIRECT"
This reverts commit 615085f6e67aec92dd225874fe135c7ed9c82323.

jamesbond3142 reported a problem on github.
Although it is uncertain that this commit is the root cause, I agree it
is highly suspicious. But there is no more way to investigate.
So let's revert it on aufs6.1..aufs6.9.

For aufs6.x-rcN (linux-v6.10-rc1), don't revert.
A similar problem may happen in the future (in aufs6.10). Hopefully
there may come another way to investigate then.

See-also: https://github.com/sfjro/aufs-standalone/issues/39