aufs: linux-v6.11, f2fs inode_lock() in file-open
By the commit in linux-v6.11,
	298b1e4182d6 2024-06-12 f2fs: fix to truncate preallocated blocks in f2fs_file_open()
f2fs acquires inode lock in file-open. For filesystems, that is
unusual behaviour.

In copy-up, aufs acquires the inode lock of the source file in order
to prevent the modification during the copy-up. It causes a
deadlock on f2fs. Simply aufs has to release the inode lock
temporarily when the source file is on f2fs.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>
(cherry picked from commit 59424f869376180af3876c3fb1efa40b97f0cc91)